JS2-1251: Add support for partly or fully readonly mapping (Ldap) backend: - configurable global readonly state on mapping SecurityEntityManager (DefaultLDAPEntityManager) - configurable allowed create/update/delete operations on entity type level - configurable allowed create/delete operations on entity relation type level
git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/portal/trunk@1101715 13f79535-47bb-0310-9956-ffa450edef68

+++ components/jetspeed-security/src/main/java/org/apache/jetspeed/security/spi/impl/JetspeedPrincipalLdapAssociationStorageManager.java
@@ -24,7 +24,7 @@ import org.apache.jetspeed.security.spi.JetspeedPrincipalAssociationStorageManag
 
 /**
  * @author <a href="mailto:vkumar@apache.org">Vivek Kumar</a>
- * @version $Id:
+ * @version $Id$
  */
 public class JetspeedPrincipalLdapAssociationStorageManager implements JetspeedPrincipalAssociationStorageManager
 {
@@ -43,10 +43,10 @@ public class JetspeedPrincipalLdapAssociationStorageManager implements JetspeedP
 
     public void addAssociation(JetspeedPrincipal from, JetspeedPrincipal to, String associationName) throws SecurityException
     {
-        if (!SynchronizationStateAccess.isSynchronizing())
+        if (!SynchronizationStateAccess.isSynchronizing() && !ldapEntityManager.isReadOnly())
         {
             SecurityEntityRelationType relationType = ldapEntityManager.getSupportedEntityRelationType(associationName, from.getType().getName(), to.getType().getName());
-            if (relationType != null)
+            if (relationType != null && relationType.isCreateAllowed())
             {
                 ldapEntityManager.addRelation(from.getName(), to.getName(), relationType);
             }
@@ -56,10 +56,10 @@ public class JetspeedPrincipalLdapAssociationStorageManager implements JetspeedP
 
     public void removeAssociation(JetspeedPrincipal from, JetspeedPrincipal to, String associationName) throws SecurityException
     {
-        if (!SynchronizationStateAccess.isSynchronizing())
+        if (!SynchronizationStateAccess.isSynchronizing() && !ldapEntityManager.isReadOnly())
         {
             SecurityEntityRelationType relationType = ldapEntityManager.getSupportedEntityRelationType(associationName, from.getType().getName(), to.getType().getName());
-            if (relationType != null)
+            if (relationType != null && relationType.isRemoveAllowed())
             {
                 ldapEntityManager.removeRelation(from.getName(), to.getName(), relationType);
             }
